+++ src/d3utils/layoutD3Utils.ts
