+++ src/mainUtils/index.ts
@@ -1,4 +1,4 @@
 export * from './databrowser';
-export * from './Parser';
+export * from './RemoteParser';
 export * from './selectioninfo';
 export * from './ToolTip';
